Split up server and DB testing code
