+++ server/test/api.edge.test.ts
@@ -1,9 +1,6 @@
-import {
-  expectDatetimeString,
-  installTestDatabaseHooks,
-  useServerForTests,
-  ndjsonParse,
-} from "./support/lib";
+import { expectDatetimeString, ndjsonParse } from "./support/lib";
+import { installTestDatabaseHooks } from "./support/database-testing";
+import { useServerForTests } from "./support/server-testing";
 import "./support/matchers";
 import { getApiKeys } from "../src/config";
 import app from "../src/app";
