+++ server/test/api.errors.test.ts
@@ -1,4 +1,5 @@
-import { installTestDatabaseHooks, useServerForTests } from "./support/lib";
+import { installTestDatabaseHooks } from "./support/database-testing";
+import { useServerForTests } from "./support/server-testing";
 import { getApiKeys } from "../src/config";
 import app from "../src/app";
 import { createLocation, updateAvailability } from "../src/db";
