+++ server/test/smart-scheduling.test.ts
@@ -1,11 +1,8 @@
 import type { AddressInfo } from "net";
 import { DateTime } from "luxon";
-import {
-  createRandomLocation,
-  useServerForTests,
-  installTestDatabaseHooks,
-  ndjsonParse,
-} from "./support/lib";
+import { createRandomLocation, ndjsonParse } from "./support/lib";
+import { installTestDatabaseHooks } from "./support/database-testing";
+import { useServerForTests } from "./support/server-testing";
 import app from "../src/app";
 import { createLocation } from "../src/db";
 import { TestLocation } from "./fixtures";
